+++ tests/recursive_learning_task7_shipgate_functional.rs
@@ -52,7 +52,7 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 
 /// Vendor-neutral fixture source for this ship-gate's `ReflectInput`
 /// builders, per PR #1175 (`validate::DEFAULT_NHI_SOURCE = "nhi"`)
-/// + PR #1174 PR9 (vendor-deflake test fixtures). The substrate now
+/// Plus PR #1174 PR9 (vendor-deflake test fixtures). The substrate now
 /// stamps `source = "nhi"` by default for any AI NHI; pinning a
 /// vendor literal here would silently couple this functional ship-gate
 /// scenario to a single vendor identity.
