+++ tests/ship_gate/grand_slam_recursive_learning.rs
@@ -75,7 +75,7 @@ use tempfile::TempDir;
 
 /// Vendor-neutral fixture source for this ship-gate's `ReflectInput`
 /// builders, per PR #1175 (`validate::DEFAULT_NHI_SOURCE = "nhi"`)
-/// + PR #1174 PR9 (vendor-deflake test fixtures). The substrate now
+/// Plus PR #1174 PR9 (vendor-deflake test fixtures). The substrate now
 /// stamps `source = "nhi"` by default for any AI NHI; pinning a
 /// vendor literal here would silently couple this ship-gate scenario
 /// to a single vendor identity.
